requirements.txt
alembic==1.17.1
bcrypt==5.0.0
blinker==1.9.0
cachelib==0.13.0
click==8.3.0
Flask==3.1.2
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
msgspec==0.19.0
psycopg2-binary==2.9.11
python-dotenv==1.2.1
SQLAlchemy==2.0.44
typing_extensions==4.15.0
Werkzeug==3.1.3
gunicorn
